Add rules for refactoring
There is much overlap in color and font styles between the various
elements.

To alleviate this, rules have been added in this commit such that each
rule is only responsible for modifying a single property. This should
hopefully reduce duplicate declarations and having the need to figure
intersecting rule groupings.